Lab1/src/requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.16
aniso8601==9.0.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.4.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
beautifulsoup4==4.12.3
black==19.10b0
boto3==1.34.63
botocore==1.34.63
braceexpand==0.1.7
cdifflib==1.2.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.0.2
clip==0.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.9
cytoolz==0.12.3
datasets==2.18.0
decorator==5.1.1
diffusers==0.27.0
dill==0.3.8
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.20.1
editdistance==0.8.1
einops==0.7.0
einops-exts==0.0.4
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.0
fasteners==0.19
fasttext==0.9.2
filelock==3.13.1
Flask==2.2.5
Flask-RESTful==0.3.10
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
future==1.0.0
g2p-en==2.1.0
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.21.4
hydra-core==1.3.2
idna==3.6
ijson==3.2.3
imageio==2.34.0
imagesize==1.4.1
importlib_metadata==7.0.2
inflect==7.0.0
iniconfig==2.0.0
intervaltree==3.1.0
ipython==8.22.2
ipywidgets==8.1.2
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.3
jiwer==2.5.2
jmespath==1.0.1
joblib==1.3.2
jupyterlab_widgets==3.0.10
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.1
latexcodec==3.0.0
lazy_loader==0.3
Levenshtein==0.22.0
lhotse==1.22.0
librosa==0.10.1
lightning-utilities==0.10.1
lilcom==1.7
llvmlite==0.42.0
loguru==0.7.2
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
megatron_core==0.5.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nemo_text_processing==0.3.0rc0
nemo_toolkit==1.23.0
nerfacc==0.5.3
networkx==3.2.1
nltk==3.8.1
numba==0.59.0
numcodecs==0.12.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.15.0
open-clip-torch==2.24.0
OpenCC==1.1.6
packaging==24.0
pandas==2.2.1
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
plac==1.4.3
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
portalocker==2.8.2
progress==1.6
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyarrow==15.0.1
pyarrow-hotfix==0.6
pybind11==2.11.1
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyloudnorm==0.1.1
PyMCubes==0.1.4
pynini==2.1.5
pyparsing==3.1.2
pypinyin==0.51.0
pypinyin-dict==0.8.0
PySocks==1.7.1
pytest==8.1.1
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.0.7
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==2.13.7
regex==2023.12.25
requests==2.31.0
resampy==0.4.3
rich==13.7.1
rouge_score==0.1.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
sacrebleu==2.4.1
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
sentencepiece==0.2.0
sentry-sdk==1.42.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
sox==1.4.1
soxr==0.3.7
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-bibtex==2.6.2
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
taming-transformers==0.0.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorstore==0.1.45
termcolor==2.4.0
text-unidecode==1.3
textdistance==4.6.1
texterrors==0.4.4
threadpoolctl==3.3.0
timm==0.9.16
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.1
torchdiffeq==0.2.3
torchmetrics==1.3.1
torchsde==0.2.6
torchvision==0.17.1
tqdm==4.66.2
traitlets==5.14.2
trampoline==0.1.2
transformers==4.38.2
trimesh==4.2.0
triton==2.2.0
typed-ast==1.5.5
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
wandb==0.16.4
wcwidth==0.2.13
webdataset==0.1.62
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.10
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
youtokentome==1.0.6
zarr==2.17.1
zipp==3.18.1
